% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqrcomp.R
\name{cqrcomp}
\alias{cqrcomp}
\title{Censored Quantile Regression comparison}
\usage{
cqrcomp(
  df1,
  df2,
  cov.names = NULL,
  delta.name = "delta",
  time.name = "y",
  taus,
  R = 100,
  verbose = F,
  method,
  bmethod = c("Bose", "jack", "xy-pair"),
  paired = F,
  test.type = c("norm2", "normInf", "bonf")
)
}
\arguments{
\item{df1}{first dataset as dataframe}

\item{df2}{second dataset as dataframe}

\item{cov.names}{covariate names in case of conditional comparison}

\item{delta.name}{for delta variable (0 censored, 1 uncensored)}

\item{time.name}{name of the time variable}

\item{taus}{values of the quantiles of interest, over which compare the quantile curves.}

\item{R}{number of bootstrap resample}

\item{verbose}{print also partial results,}

\item{method}{same as crq method, see quantreg}

\item{paired}{boolean indicating whether the samples are paried}

\item{test.type}{type of test among using norm2, normInf or bonferroni approach}
}
\value{
pcalue of the test H0: beta1(tau) = beta2(tau), for each tau in taus, using the statistics selected in test.type, and the bootstrap confidence interval specified in bmethod
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
\details{
This function loads two datasets and compare the distribution of the (conditonal) survival time
testing the equality of the quantile regression coeffcients.
The datasets can be paired or independent. In the first case, they must have the same sample sizes,
and the i-row of the first dataset is supposed to be correlated with i-row of the second.
Different statistics are possible for test: integral of the euclidean norm of the different, of the sup norm, or a Bonferroni approach testing each component separatly.
Differnt bootstrap method are available for the confidence interval of the test statistics. See quantreg for the details.
}
